services/ingestion/requirements.txt
kafka-python-ng==2.2.0
pydantic==2.5.3
python-dotenv==1.0.0
